// src/objkts/models/objkt.model.ts
export enum SortDirection {
  ASCENDING = 'asc',
  DESCENDING = 'desc',
}

export enum MarketplaceEventType {
  NULL = 'null',
  LIST_BUY = 'list_buy',
  LIST_CREATE = 'list_create',
  LIST_CANCEL = 'list_cancel',
  OFFER_CREATE = 'offer_create'
}

export enum EventType {
  NULL = 'null',
  MINT = 'mint',
  TRANSFER = 'transfer',
}